% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_list.R
\name{make_list}
\alias{make_list}
\title{Create Word Lists from Piped Text}
\usage{
make_list(
  piped,
  word_list = c(list_academic, list_business, list_dolch, list_general_plus, list_ngsl,
    list_ngsl_all, list_wiki, list_toeic)
)
}
\arguments{
\item{piped}{dataframe piped from udpipe::udpipe}

\item{word_list}{word list to use as basis for joining}
}
\value{
a dataframe of words and groups
}
\description{
Create Word Lists from Piped Text
}
