% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pos_ref}
\alias{pos_ref}
\title{udpipe upos Part-of-Speech reference}
\format{
A data frame with 14 observations and 2 variables
\describe{
  \item{upos}{form used by udpipe}
  \item{pos}{standard english}
}
}
\usage{
pos_ref
}
\description{
A dataset containing the upos and regular English for the parts of speech 
used in udpipe's upos POS column.
}
\keyword{datasets}
